… include/flecs.h
… include/flecs/private/allocator.h
… include/flecs/private/api_defines.h
… include/flecs/private/api_types.h
… include/flecs/private/sparse.h
… include/flecs/private/vec.h
… src/addons/http.c
… src/addons/rest.c
… src/addons/snapshot.c
… src/addons/stats.c
… src/bootstrap.c
… src/datastructures/allocator.c
… src/datastructures/entity_index.h
… src/datastructures/sparse.c
… src/datastructures/vec.c
… src/entity.c
… src/filter.c
… src/hierarchy.c
… src/id_record.c
… src/observable.c
… src/poly.c
… src/private_api.h
… src/private_types.h
… src/stage.c
… src/stage.h
… src/table.c
… src/table_graph.c
… src/world.c
… test/api/project.json
… test/api/src/Lookup.c