This PR adds a number of performance improvements:
- New sparse set functions that use less branching
- API operations now require entities to be alive (vs. valid), which allows them to use the faster sparse set functions
- The implementation of `ecs_has_id` has been replaced with a faster version
- The defer/defer_suspend check has been optimized
- The code to obtain mixins from poly objects has been optimized (especially `ecs_get_world`)

This PR introduces a breaking change (ids must now be alive): https://github.com/SanderMertens/flecs/discussions/466#discussioncomment-4700673

A summary of benchmarks where performance improved:

| Benchmark                           | v3.0.0      | v3.1.4      |
|-------------------------------------|-------------|-------------|
| has_empty_entity                    | 2.81ns      | 1.98ns      |
| has_id_not_found                    | 4.62ns      | 3.64ns      |
| has_id                              | 9.12ns      | 4.63ns      |
| has_16_ids                          | 9.47ns      | 4.91ns      |
| get_empty_entity                    | 2.70ns      | 1.98ns      |
| get_id_not_found                    | 6.38ns      | 3.96ns      |
| get_id                              | 7.36ns      | 4.03ns      |
| get_16_ids                          | 7.60ns      | 4.21ns      |
| get_pair                            | 7.56ns      | 4.93ns      |
| get_pair_16_targets                 | 7.69ns      | 5.17ns      |
| get_mut_id                          | 12.54ns     | 7.75ns      |
| get_mut_16_ids                      | 14.40ns     | 8.01ns      |
| set_id                              | 29.81ns     | 12.19ns     |
| set_16_ids                          | 33.45ns     | 12.71ns     |
| ref_init                            | 6.68ns      | 4.63ns      |
| ref_get                             | 2.05ns      | 2.24ns      |
| add_remove_1_tag                    | 20.90ns     | 18.43ns     |
| add_remove_2_tags                   | 27.16ns     | 23.96ns     |
| add_remove_16_tags                  | 33.31ns     | 29.42ns     |
| add_remove_1_component              | 21.59ns     | 19.55ns     |
| add_remove_2_components             | 33.59ns     | 28.10ns     |
| add_remove_16_components            | 87.08ns     | 84.78ns     |
| add_existing_1_tag                  | 11.90ns     | 10.25ns     |
| add_existing_16_tags                | 12.35ns     | 10.54ns     |
| add_existing_1_component            | 11.85ns     | 10.26ns     |
| add_existing_16_components          | 12.37ns     | 10.63ns     |
| create_delete_empty                 | 7.96ns      | 5.61ns      |
| create_delete_tree_depth_1          | 2.64us      | 1.12us      |
| create_delete_tree_depth_10         | 19.31us     | 8.51us      |
| create_delete_tree_depth_100        | 197.47us    | 88.48us     |
| create_delete_tree_depth_1000       | 2.17ms      | 908.30us    |

Find all benchmarks here: https://github.com/SanderMertens/ecs_benchmark